+++ test/chunking-form/samples/dynamic-import-dynamic/_expected/es/main.js
@@ -1,0 +1,3 @@
+var dep = 'dep';
+
+import(dep);