+++ test/chunking-form/samples/dynamic-import-dynamic/_expected/system/main.js
@@ -1,0 +1,12 @@
+System.register([], function (exports, module) {
+	'use strict';
+	return {
+		execute: function () {
+
+			var dep = 'dep';
+
+			module.import(dep);
+
+		}
+	};
+});